Backend ----------- settings.py -Added support for filtering of REST calls Customers,  Sales and Users -Added filtering mixin to the serializers and views to simplify the number of serializers Sales -Added custom autoincrementing invoice number field to Invoice model Requirements.txt -Updated requirements to include djangorestframework-queryfields

+++ backend/sales/serializers.py
@@ -1,7 +1,8 @@
-from django.contrib.auth.models import User
-
 from rest_framework import serializers
+from rest_framework_serializer_extensions.serializers import SerializerExtensionsMixin
 
+from customers.serializers import CustomerSerializer
+from users.serializers import UserSerializer
 from sales.models import (
     Quotation,
     QuotationItem,
@@ -10,23 +11,23 @@ from sales.models import (
     Invoice,
     InvoiceItem,
 )
+from inventory.models import Product
 
 
 def product_description(obj):
     return f"{obj.product.description} {obj.product.model} {obj.product.year}"
 
 
-class QuotationItemSerializer(serializers.ModelSerializer):
+class QuotationItemSerializer(SerializerExtensionsMixin, serializers.ModelSerializer):
+    class Meta:
+        model = QuotationItem
+        exclude = ["quotation"]
+
     id = serializers.CharField()
     description = serializers.SerializerMethodField()
     price = serializers.DecimalField(max_digits=16, decimal_places=2)
     vat = serializers.DecimalField(max_digits=16, decimal_places=2)
     subtotal = serializers.DecimalField(max_digits=16, decimal_places=2)
-
-    class Meta:
-        model = QuotationItem
-        # fields = "__all__"
-        exclude = ["quotation"]
 
     def create(self, *args, **kwargs):
         quotation_item = QuotationItem.objects.create(*args, **kwargs)
@@ -37,19 +38,26 @@ class QuotationItemSerializer(serializers.ModelSerializer):
         return product_description(obj)
 
 
-class QuotationSerializer(serializers.ModelSerializer):
-    items = QuotationItemSerializer(many=True)
-    user = serializers.SlugRelatedField(
-        slug_field="username", queryset=User.objects.all()
-    )
+class QuotationSerializer(SerializerExtensionsMixin, serializers.ModelSerializer):
+    class Meta:
+        model = Quotation
+        fields = "__all__"
+        expandable_fields = {
+            "items": {"serializer": QuotationItemSerializer, "many": True},
+            "user": {
+                "serializer": UserSerializer,
+                "id_source": False,
+            },
+            "customer": {"serializer": CustomerSerializer, "id_source": False},
+        }
+
+    # user = serializers.SlugRelatedField(
+    #     slug_field="username", queryset=User.objects.all()
+    # )
 
     amount = serializers.DecimalField(max_digits=16, decimal_places=2)
     vat = serializers.DecimalField(max_digits=16, decimal_places=2)
     total = serializers.DecimalField(max_digits=16, decimal_places=2)
-
-    class Meta:
-        model = Quotation
-        fields = "__all__"
 
     def create(self, validated_data):
         items = validated_data.pop("items")
@@ -77,41 +85,34 @@ class QuotationSerializer(serializers.ModelSerializer):
         instance.memo = validated_data.get("memo", instance.memo)
 
         items = validated_data.get("items")
-        item_ids = map(lambda item: item.get("id"), items)
-        item_ids = filter(lambda id: id.isnumeric(), item_ids)
-        instance.items.exclude(id__in=item_ids).delete()
-
-        for item in items:
-            id = item.get("id")
-            try:
-                id = int(id)
-                quotation_item_instance = QuotationItem.objects.get(id=id)
-                QuotationItemSerializer().update(quotation_item_instance, item)
-            except (TypeError, ValueError, QuotationItem.DoesNotExist):
-                item.pop("id")
-                QuotationItemSerializer().create(**item, quotation=instance)
+        if items:
+            item_ids = map(lambda item: item.get("id"), items)
+            item_ids = filter(lambda id: id.isnumeric(), item_ids)
+            instance.items.exclude(id__in=item_ids).delete()
+            for item in items:
+                item_id = item.get("id")
+                try:
+                    item_id = int(item_id)
+                    quotation_item_instance = QuotationItem.objects.get(id=item_id)
+                    QuotationItemSerializer().update(quotation_item_instance, item)
+                except (TypeError, ValueError, QuotationItem.DoesNotExist):
+                    item.pop("id")
+                    QuotationItemSerializer().create(**item, quotation=instance)
 
         instance.save()
         return instance
 
 
-class QuotationListSerializer(serializers.ModelSerializer):
+class OrderItemSerializer(SerializerExtensionsMixin, serializers.ModelSerializer):
     class Meta:
-        model = Quotation
-        fields = ["reference_number"]
+        model = OrderItem
+        exclude = ["order"]
 
-
-class OrderItemSerializer(serializers.ModelSerializer):
     id = serializers.CharField()
     description = serializers.SerializerMethodField()
     price = serializers.DecimalField(max_digits=16, decimal_places=2)
     vat = serializers.DecimalField(max_digits=16, decimal_places=2)
     subtotal = serializers.DecimalField(max_digits=16, decimal_places=2)
-
-    class Meta:
-        model = OrderItem
-        # fields = "__all__"
-        exclude = ["order"]
 
     def create(self, *args, **kwargs):
         order_item = OrderItem.objects.create(*args, **kwargs)
@@ -122,19 +123,22 @@ class OrderItemSerializer(serializers.ModelSerializer):
         return product_description(obj)
 
 
-class OrderSerializer(serializers.ModelSerializer):
-    items = OrderItemSerializer(many=True)
-    user = serializers.SlugRelatedField(
-        slug_field="username", queryset=User.objects.all()
-    )
+class OrderSerializer(SerializerExtensionsMixin, serializers.ModelSerializer):
+    class Meta:
+        model = Order
+        fields = "__all__"
+        expandable_fields = {
+            "items": {"serializer": OrderItemSerializer, "many": True},
+            "user": {
+                "serializer": UserSerializer,
+                "id_source": False,
+            },
+            "customer": {"serializer": CustomerSerializer, "id_source": False},
+        }
 
     amount = serializers.DecimalField(max_digits=16, decimal_places=2)
     vat = serializers.DecimalField(max_digits=16, decimal_places=2)
     total = serializers.DecimalField(max_digits=16, decimal_places=2)
-
-    class Meta:
-        model = Order
-        fields = "__all__"
 
     def create(self, validated_data):
         items = validated_data.pop("items")
@@ -163,37 +167,126 @@ class OrderSerializer(serializers.ModelSerializer):
         instance.memo = validated_data.get("memo", instance.memo)
 
         items = validated_data.get("items")
-        item_ids = map(lambda item: item.get("id"), items)
-        item_ids = filter(lambda id: id.isnumeric(), item_ids)
-        instance.items.exclude(id__in=item_ids).delete()
+        if items:
+            item_ids = map(lambda item: item.get("id"), items)
+            item_ids = filter(lambda id: id.isnumeric(), item_ids)
+            instance.items.exclude(id__in=item_ids).delete()
 
-        for item in items:
-            id = item.get("id")
-            try:
-                id = int(id)
-                order_item_instance = OrderItem.objects.get(id=id)
-                OrderItemSerializer().update(order_item_instance, item)
-            except (TypeError, ValueError, OrderItem.DoesNotExist):
-                item.pop("id")
-                OrderItemSerializer().create(**item, order=instance)
+            for item in items:
+                item_id = item.get("id")
+                try:
+                    item_id = int(item_id)
+                    order_item_instance = OrderItem.objects.get(id=item_id)
+                    OrderItemSerializer().update(order_item_instance, item)
+                except (TypeError, ValueError, OrderItem.DoesNotExist):
+                    item.pop("id")
+                    OrderItemSerializer().create(**item, order=instance)
 
         instance.save()
         return instance
 
 
-class OrderListSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Order
-        fields = ["reference_number"]
-
-
-class InvoiceItemSerializer(serializers.ModelSerializer):
+class InvoiceItemSerializer(SerializerExtensionsMixin, serializers.ModelSerializer):
     class Meta:
         model = InvoiceItem
-        fields = "__all__"
+        exclude = ["invoice"]
+
+    id = serializers.CharField()
+    description = serializers.SerializerMethodField()
+    price = serializers.DecimalField(max_digits=16, decimal_places=2)
+    vat = serializers.DecimalField(max_digits=16, decimal_places=2)
+    subtotal = serializers.DecimalField(max_digits=16, decimal_places=2)
+
+    def create(self, *args, **kwargs):
+        invoice_item = InvoiceItem.objects.create(*args, **kwargs)
+
+        return invoice_item
+
+    def get_description(self, obj):
+        return product_description(obj)
 
 
-class InvoiceSerializer(serializers.ModelSerializer):
+class InvoiceSerializer(SerializerExtensionsMixin, serializers.ModelSerializer):
     class Meta:
         model = Invoice
         fields = "__all__"
+        expandable_fields = {
+            "items": {"serializer": InvoiceItemSerializer, "many": True},
+            "user": {
+                "serializer": UserSerializer,
+                "id_source": False,
+            },
+            "customer": {"serializer": CustomerSerializer, "id_source": False},
+        }
+
+    amount = serializers.DecimalField(max_digits=16, decimal_places=2)
+    vat = serializers.DecimalField(max_digits=16, decimal_places=2)
+    total = serializers.DecimalField(max_digits=16, decimal_places=2)
+
+    def create(self, validated_data):
+        items = validated_data.pop("items")
+
+        invoice = Invoice.objects.create(**validated_data)
+
+        for item in items:
+            try:
+                item.pop("id")
+            except KeyError:
+                print("Invoice item does not have id field.")
+            finally:
+                InvoiceItem.objects.create(invoice=invoice, **item)
+
+        invoice.save()
+        return invoice
+
+    def update(self, instance, validated_data):
+        instance.user = validated_data.get("user", instance.user)
+        instance.customer = validated_data.get("customer", instance.customer)
+
+        instance.amount = validated_data.get("amount", instance.amount)
+        instance.vat = validated_data.get("vat", instance.vat)
+        instance.total = validated_data.get("total", instance.total)
+
+        instance.memo = validated_data.get("memo", instance.memo)
+
+        items = validated_data.get("items")
+
+        item_ids = map(lambda item: item.get("id"), items)
+        item_ids = list(filter(lambda id: id.isnumeric(), item_ids))
+        removed_rows = instance.items.exclude(id__in=item_ids)
+        for removed_row in removed_rows:
+            product_instance = Product.objects.get(product_number=removed_row.product)
+            product_instance.quantity += removed_row.quantity
+            product_instance.save()
+        removed_rows.delete()
+
+        for item in items:
+            item_id = item.get("id")
+            try:
+                item_id = int(item_id)
+                invoice_item_instance = InvoiceItem.objects.get(id=item_id)
+                old_quantity = invoice_item_instance.quantity
+
+                InvoiceItemSerializer().update(invoice_item_instance, item)
+
+                product_number = item.pop("product")
+                new_quantity = item.pop("quantity")
+                quantity_difference = new_quantity - old_quantity
+
+                product_instance = Product.objects.get(product_number=product_number)
+                product_instance.quantity -= quantity_difference
+                product_instance.save()
+
+            except (TypeError, ValueError, OrderItem.DoesNotExist):
+                item.pop("id")
+                InvoiceItemSerializer().create(**item, invoice=instance)
+
+                product_number = item.pop("product")
+                new_quantity = item.pop("quantity")
+
+                product_instance = Product.objects.get(product_number=product_number)
+                product_instance.quantity -= new_quantity
+                product_instance.save()
+
+        instance.save()
+        return instance
